JRMP2-1 Check and fix file module (relocated some function)
